Unify QueryNotifier and RouteNotifier into PerformanceNotifier
The public interface remains the same:

* Airbrake.notify_request
* Airbrake.notify_query

What's different is how we implement this behind the
scenes. `PerformanceNotifier` uses duck typing and accepts `Request` and `Query`
Structs that implement a contract. These Structs must implement `#name` and
`#to_h`.

`#name` is used to distinguish payload type. This is needed so
PerformanceNotifier can `#notify` payload of different types during one flush
cycle. Upon flushing the payload will be grouped and sent to appropriate
endpoints. It means that a notifier can send multiple requests during one cycle.
